+++ imod/formats/gen/__init__.py
@@ -1,0 +1,1 @@
+from imod.formats.gen.gen import read, write